add topic_category module
